Removed proof of concept
